Refine functionalities and update java doc.
