+++ src/test/java/org/datasyslab/geospark/babylon/ChoroplethmapTest.java
@@ -1,23 +1,31 @@
 /**
- * FILE: ParallelVisualizationTest.java
- * PATH: org.datasyslab.babylon.ParallelVisualizationTest.java
+ * FILE: ChoroplethmapTest.java
+ * PATH: org.datasyslab.geospark.babylon.ChoroplethmapTest.java
  * Copyright (c) 2017 Arizona State University Data Systems Lab
  * All rights reserved.
  */
-package org.datasyslab.babylon;
+package org.datasyslab.geospark.babylon;
 
+import java.awt.Color;
 import java.io.InputStream;
+import java.io.Serializable;
 import java.util.Properties;
 
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 import org.apache.spark.SparkConf;
+import org.apache.spark.api.java.JavaPairRDD;
 import org.apache.spark.api.java.JavaSparkContext;
+import org.datasyslab.babylon.core.OverlayOperator;
 import org.datasyslab.babylon.extension.imageGenerator.NativeJavaImageGenerator;
-import org.datasyslab.babylon.extension.visualizationEffect.HeatMap;
+import org.datasyslab.babylon.extension.visualizationEffect.ChoroplethMap;
+import org.datasyslab.babylon.extension.visualizationEffect.ScatterPlot;
 import org.datasyslab.babylon.utils.ImageType;
 import org.datasyslab.geospark.enums.FileDataSplitter;
-import org.datasyslab.geospark.spatialRDD.LineStringRDD;
+import org.datasyslab.geospark.enums.GridType;
+import org.datasyslab.geospark.enums.IndexType;
+import org.datasyslab.geospark.showcase.UserSuppliedPolygonMapper;
+import org.datasyslab.geospark.spatialOperator.JoinQuery;
 import org.datasyslab.geospark.spatialRDD.PointRDD;
 import org.datasyslab.geospark.spatialRDD.PolygonRDD;
 import org.datasyslab.geospark.spatialRDD.RectangleRDD;
@@ -26,12 +34,13 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import com.vividsolutions.jts.geom.Envelope;
+import com.vividsolutions.jts.geom.Polygon;
 
 /**
- * The Class ParallelVisualizationTest.
+ * The Class ChoroplethmapTest.
  */
-public class ParallelVisualizationTest {
-
+public class ChoroplethmapTest implements Serializable{
+	
     /** The spark context. */
     static JavaSparkContext sparkContext;
     
@@ -92,18 +101,6 @@ public class ParallelVisualizationTest {
     /** The US main land boundary. */
     static Envelope USMainLandBoundary;
     
-    /** The resolution X. */
-    static int resolutionX;
-    
-    /** The resolution Y. */
-    static int resolutionY;
-    
-    /** The partition X. */
-    static int partitionX;
-    
-    /** The partition Y. */
-    static int partitionY;
-	
 	/**
 	 * Sets the up before class.
 	 *
@@ -111,49 +108,41 @@ public class ParallelVisualizationTest {
 	 */
 	@BeforeClass
 	public static void setUpBeforeClass() throws Exception {
-		SparkConf sparkConf = new SparkConf().setAppName("ParallelVisualizationTest").setMaster("local[4]");
+		SparkConf sparkConf = new SparkConf().setAppName("ChoroplethmapTest").setMaster("local[4]");
 		sparkContext = new JavaSparkContext(sparkConf);
         Logger.getLogger("org").setLevel(Level.WARN);
         Logger.getLogger("akka").setLevel(Level.WARN);
         prop = new Properties();
         
-        inputProp = ParallelVisualizationTest.class.getClassLoader().getResourceAsStream("babylon.point.properties");
+        inputProp = ChoroplethmapTest.class.getClassLoader().getResourceAsStream("babylon.point.properties");
         prop.load(inputProp);
-        PointInputLocation = "file://"+ParallelVisualizationTest.class.getClassLoader().getResource(prop.getProperty("inputLocation")).getPath();
+        PointInputLocation = "file://"+ChoroplethmapTest.class.getClassLoader().getResource(prop.getProperty("inputLocation")).getPath();
         PointOffset = Integer.parseInt(prop.getProperty("offset"));;
         PointSplitter = FileDataSplitter.getFileDataSplitter(prop.getProperty("splitter"));
         PointNumPartitions = Integer.parseInt(prop.getProperty("numPartitions"));
         
-        inputProp = ParallelVisualizationTest.class.getClassLoader().getResourceAsStream("babylon.rectangle.properties");
+        inputProp = ChoroplethmapTest.class.getClassLoader().getResourceAsStream("babylon.rectangle.properties");
         prop.load(inputProp);
-        RectangleInputLocation = "file://"+ParallelVisualizationTest.class.getClassLoader().getResource(prop.getProperty("inputLocation")).getPath();
+        RectangleInputLocation = "file://"+ChoroplethmapTest.class.getClassLoader().getResource(prop.getProperty("inputLocation")).getPath();
         RectangleOffset = Integer.parseInt(prop.getProperty("offset"));
         RectangleSplitter = FileDataSplitter.getFileDataSplitter(prop.getProperty("splitter"));
         RectangleNumPartitions = Integer.parseInt(prop.getProperty("numPartitions"));
         
-        inputProp = ParallelVisualizationTest.class.getClassLoader().getResourceAsStream("babylon.polygon.properties");
+        inputProp = ChoroplethmapTest.class.getClassLoader().getResourceAsStream("babylon.polygon.properties");
         prop.load(inputProp);
-        PolygonInputLocation = "file://"+ParallelVisualizationTest.class.getClassLoader().getResource(prop.getProperty("inputLocation")).getPath();
+        PolygonInputLocation = "file://"+ChoroplethmapTest.class.getClassLoader().getResource(prop.getProperty("inputLocation")).getPath();
         PolygonOffset = Integer.parseInt(prop.getProperty("offset"));
         PolygonSplitter = FileDataSplitter.getFileDataSplitter(prop.getProperty("splitter"));
         PolygonNumPartitions = Integer.parseInt(prop.getProperty("numPartitions"));
 
-        inputProp = ParallelVisualizationTest.class.getClassLoader().getResourceAsStream("babylon.linestring.properties");
+        inputProp = ChoroplethmapTest.class.getClassLoader().getResourceAsStream("babylon.linestring.properties");
         prop.load(inputProp);
-        LineStringInputLocation = "file://"+ParallelVisualizationTest.class.getClassLoader().getResource(prop.getProperty("inputLocation")).getPath();
+        LineStringInputLocation = "file://"+ChoroplethmapTest.class.getClassLoader().getResource(prop.getProperty("inputLocation")).getPath();
         LineStringOffset = Integer.parseInt(prop.getProperty("offset"));
         LineStringSplitter = FileDataSplitter.getFileDataSplitter(prop.getProperty("splitter"));
         LineStringNumPartitions = Integer.parseInt(prop.getProperty("numPartitions"));
         
         USMainLandBoundary = new Envelope(-126.790180,-64.630926,24.863836,50.000);
-        
-        resolutionX = 1000;
-        
-        resolutionY = 600;
-        
-        partitionX = 2;
-        
-        partitionY = 2;
 	}
 
 	/**
@@ -167,33 +156,33 @@ public class ParallelVisualizationTest {
 	}
 
 	/**
-	 * Test point RDD visualization.
-	 *
-	 * @throws Exception the exception
-	 */
-	@Test
-	public void testPointRDDVisualization() throws Exception {
-		PointRDD spatialRDD = new PointRDD(sparkContext, PointInputLocation, PointOffset, PointSplitter, false, PointNumPartitions);
-		HeatMap visualizationOperator = new HeatMap(resolutionX,resolutionY,USMainLandBoundary,false,2,partitionX,partitionY,true,true);
-		visualizationOperator.Visualize(sparkContext, spatialRDD);
-		visualizationOperator.stitchImagePartitions();
-		NativeJavaImageGenerator imageGenerator = new  NativeJavaImageGenerator();
-		imageGenerator.SaveAsFile(visualizationOperator.pixelImage, "./target/parallelvisualization/PointRDD",ImageType.gif);
-	}
-	
-	/**
 	 * Test rectangle RDD visualization.
 	 *
 	 * @throws Exception the exception
 	 */
 	@Test
 	public void testRectangleRDDVisualization() throws Exception {
-		RectangleRDD spatialRDD = new RectangleRDD(sparkContext, RectangleInputLocation, RectangleSplitter, false, RectangleNumPartitions);
-		HeatMap visualizationOperator = new HeatMap(resolutionX,resolutionY,USMainLandBoundary,false,2,partitionX,partitionY,true,true);
-		visualizationOperator.Visualize(sparkContext, spatialRDD);
-		visualizationOperator.stitchImagePartitions();
-		NativeJavaImageGenerator imageGenerator = new  NativeJavaImageGenerator();
-		imageGenerator.SaveAsFile(visualizationOperator.pixelImage, "./target/parallelvisualization/RectangleRDD",ImageType.png);
+		PointRDD spatialRDD = new PointRDD(sparkContext, PointInputLocation, PointOffset, PointSplitter, false, PointNumPartitions);
+		RectangleRDD queryRDD = new RectangleRDD(sparkContext, RectangleInputLocation, RectangleSplitter, false, RectangleNumPartitions);
+		spatialRDD.spatialPartitioning(GridType.RTREE);
+		queryRDD.spatialPartitioning(spatialRDD.grids);
+  		spatialRDD.buildIndex(IndexType.RTREE,true);
+  		JavaPairRDD<Envelope,Long> joinResult = JoinQuery.SpatialJoinQueryCountByKey(spatialRDD,queryRDD,true);
+  		
+  		ChoroplethMap visualizationOperator = new ChoroplethMap(1000,600,USMainLandBoundary,false);
+		visualizationOperator.CustomizeColor(255, 255, 255, 255, Color.RED, true);
+		visualizationOperator.Visualize(sparkContext, joinResult);
+		
+		ScatterPlot frontImage = new ScatterPlot(1000,600,USMainLandBoundary,false);
+		frontImage.CustomizeColor(0, 0, 0, 255, Color.GREEN, true);
+		frontImage.Visualize(sparkContext, queryRDD);
+
+		
+		OverlayOperator overlayOperator = new OverlayOperator(visualizationOperator.pixelImage);
+		overlayOperator.JoinImage(frontImage.pixelImage);
+		
+		NativeJavaImageGenerator imageGenerator = new NativeJavaImageGenerator();
+		imageGenerator.SaveAsFile(overlayOperator.backImage,"./target/choroplethmap/RectangleRDD-combined", ImageType.PNG);
 	}
 	
 	/**
@@ -204,26 +193,27 @@ public class ParallelVisualizationTest {
 	@Test
 	public void testPolygonRDDVisualization() throws Exception {
 		//UserSuppliedPolygonMapper userSuppliedPolygonMapper = new UserSuppliedPolygonMapper();
-		PolygonRDD spatialRDD = new PolygonRDD(sparkContext, PolygonInputLocation, PolygonSplitter, false, PolygonNumPartitions);
-		HeatMap visualizationOperator = new HeatMap(resolutionX,resolutionY,USMainLandBoundary,false,2,partitionX,partitionY,true,true);
-		visualizationOperator.Visualize(sparkContext, spatialRDD);
-		visualizationOperator.stitchImagePartitions();
-		NativeJavaImageGenerator imageGenerator = new  NativeJavaImageGenerator();
-		imageGenerator.SaveAsFile(visualizationOperator.pixelImage, "./target/parallelvisualization/PolygonRDD",ImageType.jpg);
+		PointRDD spatialRDD = new PointRDD(sparkContext, PointInputLocation, PointOffset, PointSplitter, false, PointNumPartitions);
+		PolygonRDD queryRDD = new PolygonRDD(sparkContext, PolygonInputLocation,  PolygonSplitter, false, PolygonNumPartitions);
+		spatialRDD.spatialPartitioning(GridType.RTREE);
+		queryRDD.spatialPartitioning(spatialRDD.grids);
+  		spatialRDD.buildIndex(IndexType.RTREE,true);
+		JavaPairRDD<Polygon,Long> joinResult = JoinQuery.SpatialJoinQueryCountByKey(spatialRDD,queryRDD,true);
+
+		ChoroplethMap visualizationOperator = new ChoroplethMap(1000,600,USMainLandBoundary,false);
+		visualizationOperator.CustomizeColor(255, 255, 255, 255, Color.RED, true);
+		visualizationOperator.Visualize(sparkContext, joinResult);
+		
+		ScatterPlot frontImage = new ScatterPlot(1000,600,USMainLandBoundary,false);
+		frontImage.CustomizeColor(0, 0, 0, 255, Color.GREEN, true);
+		frontImage.Visualize(sparkContext, queryRDD);
+		
+		OverlayOperator overlayOperator = new OverlayOperator(visualizationOperator.pixelImage);
+		overlayOperator.JoinImage(frontImage.pixelImage);
+		
+		NativeJavaImageGenerator imageGenerator = new NativeJavaImageGenerator();
+		imageGenerator.SaveAsFile(overlayOperator.backImage, "./target/choroplethmap/PolygonRDD-combined", ImageType.GIF);
+		
 	}
-	
-	/**
-	 * Test line string RDD visualization.
-	 *
-	 * @throws Exception the exception
-	 */
-	@Test
-	public void testLineStringRDDVisualization() throws Exception {
-		LineStringRDD spatialRDD = new LineStringRDD(sparkContext, LineStringInputLocation, LineStringSplitter, false, LineStringNumPartitions);
-		HeatMap visualizationOperator = new HeatMap(resolutionX,resolutionY,USMainLandBoundary,false,2,partitionX,partitionY,true,true);
-		visualizationOperator.Visualize(sparkContext, spatialRDD);
-		visualizationOperator.stitchImagePartitions();
-		NativeJavaImageGenerator imageGenerator = new  NativeJavaImageGenerator();
-		imageGenerator.SaveAsFile(visualizationOperator.pixelImage, "./target/parallelvisualization/LineStringRDD",ImageType.gif);
-	}
+
 }
