+++ src/test/java/org/datasyslab/geospark/geometryObjects/CircleTest.java
@@ -10,12 +10,6 @@ import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
-/**
- * 
- * @author Arizona State University DataSystems Lab
- *
- */
-
 import com.vividsolutions.jts.geom.Coordinate;
 import com.vividsolutions.jts.geom.Envelope;
 import com.vividsolutions.jts.geom.GeometryFactory;
